# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FeedSystem Video backend API documentation.",
        "title": "FeedSystem Video API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/feed/listByFollowing": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "List videos from followed vloggers",
                "parameters": [
                    {
                        "description": "following feed payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feed.ListByFollowingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.ListByFollowingResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feed/listByPopularity": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "List videos by popularity",
                "parameters": [
                    {
                        "description": "popularity feed payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feed.ListByPopularityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.ListByPopularityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feed/listLatest": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "List latest videos",
                "parameters": [
                    {
                        "description": "latest feed payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feed.ListLatestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.ListLatestResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feed/listLikesCount": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "List videos by likes count",
                "parameters": [
                    {
                        "description": "likes ranking payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/feed.ListLikesCountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/feed.ListLikesCountResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/swagger.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "feed.FeedAuthor": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "feed.FeedVideoItem": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/feed.FeedAuthor"
                },
                "cover_url": {
                    "type": "string"
                },
                "create_time": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "likes_count": {
                    "type": "integer"
                },
                "play_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "feed.ListByFollowingRequest": {
            "type": "object",
            "properties": {
                "latest_time": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "feed.ListByFollowingResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "next_time": {
                    "type": "integer"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedVideoItem"
                    }
                }
            }
        },
        "feed.ListByPopularityRequest": {
            "type": "object",
            "properties": {
                "as_of": {
                    "description": "服务器返回的分钟时间戳；第一页传0",
                    "type": "integer"
                },
                "latest_before": {
                    "type": "string"
                },
                "latest_id_before": {
                    "type": "integer"
                },
                "latest_popularity": {
                    "description": "DB fallback 用（可选）",
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "description": "下一页从这里开始；第一页传0",
                    "type": "integer"
                }
            }
        },
        "feed.ListByPopularityResponse": {
            "type": "object",
            "properties": {
                "as_of": {
                    "type": "integer"
                },
                "has_more": {
                    "type": "boolean"
                },
                "next_latest_before": {
                    "type": "string"
                },
                "next_latest_id_before": {
                    "type": "integer"
                },
                "next_latest_popularity": {
                    "type": "integer"
                },
                "next_offset": {
                    "type": "integer"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedVideoItem"
                    }
                }
            }
        },
        "feed.ListLatestRequest": {
            "type": "object",
            "properties": {
                "latest_time": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "feed.ListLatestResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "next_time": {
                    "type": "integer"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedVideoItem"
                    }
                }
            }
        },
        "feed.ListLikesCountRequest": {
            "type": "object",
            "properties": {
                "id_before": {
                    "type": "integer"
                },
                "likes_count_before": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                }
            }
        },
        "feed.ListLikesCountResponse": {
            "type": "object",
            "properties": {
                "has_more": {
                    "type": "boolean"
                },
                "next_id_before": {
                    "type": "integer"
                },
                "next_likes_count_before": {
                    "type": "integer"
                },
                "video_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/feed.FeedVideoItem"
                    }
                }
            }
        },
...
        "swagger.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}